Probtraj results now in ProbTrajResult
